Refactoring; + brought Media container result more in line with how other results are handled
— Namely, it it now returns yes/no if there’s media in the
getStringResult() method, instead of overriding the whole method for
assembling the text json
