+++ app/src/main/java/com/kpit/chhotescientists/model/TextQuestion.java
@@ -1,13 +1,11 @@
 package com.kpit.chhotescientists.model;
 
 import android.app.Activity;
-import android.content.Context;
 import android.os.Parcel;
-import android.view.View;
 import android.widget.EditText;
 
-import com.kpit.chhotescientists.model.result_views.ResultEditTextContainer;
-import com.kpit.chhotescientists.model.result_views.ResultViewContainer;
+import com.kpit.chhotescientists.model.view_containers.ResultEditTextContainer;
+import com.kpit.chhotescientists.model.view_containers.ResultViewContainer;
 
 /**
  * See parent CheckInQuestion for documentation.
